requirements.txt
mail-parser
requests
python-dotenv
spamassassin-client
streamlit
matplotlib
pandas
